Refactors code
Description:
- Creates shouldUseSecuredRegistryConnection method in AASServerComponent
- Refactors code
- Modifies TCs in TestAASServerWithSecuredRegistry
