+++ basyx.components/basyx.components.docker/basyx.components.AASServer/src/main/resources/context.properties
@@ -39,4 +39,4 @@ contextPort=4001
 # CORS can be configured by defining the origin 
 # Use wildcard (*) to indicate that the requests from all domains are allowed 
 
-# accessControlAllowOrigin=http://www.example.com
+# accessControlAllowOrigin=http://www.example.com
